fashion_rag/requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
pandas
fastapi
uvicorn
streamlit
redis
python-dotenv
langsmith
pydantic
httpx
